+++ next-app/src/components/views/MyPage/List/PostList/index.ts
@@ -1,0 +1,1 @@
+export { default } from './PostList'
